Added nats, which are passed by reference. See the test_nat and test_nat_ops .ank and .cpp files for example usage. Currently, only addition, subtraction, and comparisons are allowed on nats; however, I will add bitwise operations and modular multiplication soon. All operations are constant-time, assuming they are also constant-time on uint32_ts.
I also added a set_max_nat_bits(n) function to allow for arbitrary precision. It can only be called before any nats are defined, and effectively sets the computation speed and precision. n must be an integer >= 32 and n must be divisible by 32.

There are LOTS of changes to the code to allow for nat support---my recommendation would not be to attempt a merge with an older version.